moe: rename small-M Gluon routing to gluon_decode_routing_gfx950
Rename the fused small-M routing kernel module and its test to
gluon_decode_routing_gfx950 to make the decode-path target explicit,
and move the test under test/ops/moe/ to mirror the source package
layout. Imports updated; no behavior change.
